+++ tiledesk-chatbot-util/CHANGELOG.md
@@ -2,6 +2,12 @@
 
 This library is on npm: https://www.npmjs.com/package/@tiledesk/tiledesk-chatbot-util
 
+### 0.8.27
+- added findDirective() method to parse_directives() return object
+
+### 0.8.26
+- added parse_directives()
+
 ### 0.8.25
 - added test for findSplits(): no splits will return "one" command with the message text content